+++ classes/Event/Deleted.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Neat\Object\Event;
+
+use Neat\Object\Event;
+
+class Deleted extends Event
+{
+    use Entity;
+}