+++ classes/Event/Deleting.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Neat\Object\Event;
+
+use Neat\Object\Event;
+
+class Deleting extends Event
+{
+    use Entity;
+}