+++ classes/Event/Entity.php
@@ -1,0 +1,27 @@
+<?php
+
+namespace Neat\Object\Event;
+
+trait Entity
+{
+    /** @var object */
+    private $entity;
+
+    /**
+     * Event constructor
+     *
+     * @param object $entity
+     */
+    public function __construct($entity)
+    {
+        $this->entity = $entity;
+    }
+
+    /**
+     * @return object
+     */
+    public function entity()
+    {
+        return $this->entity;
+    }
+}