+++ classes/Event/Loaded.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Neat\Object\Event;
+
+use Neat\Object\Event;
+
+class Loaded extends Event
+{
+    use Entity;
+}