+++ classes/Event/Loading.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Neat\Object\Event;
+
+use Neat\Object\Event;
+
+class Loading extends Event
+{
+    use Entity;
+}