+++ classes/Event/Storing.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Neat\Object\Event;
+
+use Neat\Object\Event;
+
+class Storing extends Event
+{
+    use Entity;
+}