+++ tests/Helper/Events.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace Neat\Object\Test\Helper;
+
+use Neat\Object\Event;
+
+class Events
+{
+    const EVENTS = [
+        Event::LOADING  => Event\Loading::class,
+        Event::LOADED   => Event\Loaded::class,
+        Event::STORING  => Event\Storing::class,
+        Event::STORED   => Event\Stored::class,
+        Event::DELETING => Event\Deleting::class,
+        Event::DELETED  => Event\Deleted::class,
+    ];
+}